Test improvements
Move tests to IntegrationTests folder + small cleanup in tests
